+++ frontend/src/app/components/StudentInfoInput.tsx
